refactor code/folder structure
