--- backend/src/core/ocel1_to_ocel2_converter/mod.rs
@@ -1,2 +1,0 @@
-pub mod utils;
-pub mod converter;
